format time

+++ lib/accuweather/hourly/accu_hourly_data_entity.dart
@@ -5,7 +5,7 @@ class AccuHourlyDataEntity with JsonConvert<AccuHourlyDataEntity> {
 	@JSONField(name: "DateTime")
 	String dateTime;
 	@JSONField(name: "EpochDateTime")
-	double epochDateTime;
+	int epochDateTime;
 	@JSONField(name: "WeatherIcon")
 	double weatherIcon;
 	@JSONField(name: "IconPhrase")
